requirements.txt
numpy
pandas
matplotlib
streamlit
opencv_python
opencv-python-headless
Pillow